feat: Handle different OCIO configs in adapter
