+++ src/deadline/nuke_adaptor/NukeClient/nuke_client.py
@@ -117,7 +117,7 @@ class NukeClient(_ClientInterface):
 
     def _map_ocio_config(self):
         """If the OCIO config contains absolute search paths, apply path mapping rules and create a new config"""
-        ocio_config_path = nuke_ocio.get_custom_config_path()
+        ocio_config_path = nuke_ocio.get_ocio_config_path()
         ocio_config = nuke_ocio.create_config_from_file(ocio_config_path)
         if nuke_ocio.config_has_absolute_search_paths(ocio_config):
             # make all search paths absolute since the new config will be saved in the nuke temp dir
